Template setup
Template setup for user
