+++ app/Http/Controllers/Dashboard/DashboardController.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace App\Http\Controllers\Dashboard;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+class DashboardController extends Controller
+{
+   public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
+    /**
+     * Show the application dashboard.
+     *
+     * @return \Illuminate\Contracts\Support\Renderable
+     */
+    public function index()
+    {
+        return view('main.pages.dashboard');
+    }
+
+    public function body()
+    {
+        return view('main.pages.dashboard');
+    }
+}
