+++ app/Providers/RouteServiceProvider.php
@@ -17,7 +17,7 @@ class RouteServiceProvider extends ServiceProvider
      *
      * @var string
      */
-    public const HOME = '/Home';
+    public const HOME = '/dashboard';
 
     /**
      * The controller namespace for the application.
