+++ resources/views/main/components/footer.blade.php
@@ -1,0 +1,11 @@
+<!-- FOOTER -->
+			<footer class="footer">
+				<div class="container">
+					<div class="row align-items-center flex-row-reverse">
+						<div class="col-md-12 col-sm-12 text-center">
+							Copyright © 2019 <a href="#">Volgh</a>. Designed by <a href="#"> Spruko Technologies Pvt.Ltd </a> All rights reserved.
+						</div>
+					</div>
+				</div>
+			</footer>
+<!-- FOOTER END -->
