+++ resources/views/main/components/left-bar.blade.php
@@ -1,0 +1,407 @@
+		<!-- Start Switcher -->
+		<div class="switcher-wrapper">
+			<div class="demo_changer">
+				<div class="demo-icon bg_dark"><i class="fa fa-cog fa-spin  text_primary"></i></div>
+				<div class="form_holder sidebar-right1">
+					<div class="row">
+						<div class="predefined_styles">
+							<div class="swichermainleft border-bottom text-center">
+								<div class="p-3">
+									<a href="https://codecanyon.net/user/sprukosoft/portfolio" class="btn btn-danger btn-block mt-0">View Demo</a>
+									<a href="https://codecanyon.net/item/volgh-laravel-admin-template/25860093" class="btn btn-success btn-block">Buy Now</a>
+									<a href="https://themeforest.net/user/sprukosoft/portfolio" class="btn btn-info btn-block">Our Portfolio</a>
+								</div>
+							</div>
+							<div class="swichermainleft">
+								<h4>Navigation Style</h4>
+								<div class="pl-3 pr-3">
+									<a class="btn btn-primary btn-block" href="https://laravel.spruko.com/volgh/horizontal_light/index">
+										Horizontal
+									</a>
+									<a class="btn btn-secondary btn-block" href="index.html">
+										Left-menu
+									</a>
+								</div>
+							</div>
+							<div class="swichermainleft">
+								<h4>Skin Modes</h4>
+								<div class="switch_section">
+									<div class="switch-toggle d-flex">
+										<span class="mr-auto">Light Mode</span>
+										<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch3" class="onoffswitch2-checkbox" checked>
+											<label for="myonoffswitch3" class="onoffswitch2-label"></label>
+										</div>
+									</div>
+									<div class="switch-toggle d-flex">
+										<span class="mr-auto">Transparent Mode</span>
+										<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch4" class="onoffswitch2-checkbox">
+											<label for="myonoffswitch4" class="onoffswitch2-label"></label>
+										</div>
+									</div>
+									<div class="switch-toggle d-flex">
+										<span class="mr-auto">Dark Mode</span>
+										<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch16" class="onoffswitch2-checkbox">
+											<label for="myonoffswitch16" class="onoffswitch2-label"></label>
+										</div>
+									</div>
+								</div>
+							</div>
+							<div class="clearfix"></div>
+							<div class="swichermainleft">
+								<div class="switch_bg-img">
+									<div class=" d-flex">
+										<span class="mr-auto">Left-menu Bg-Image</span>
+										<div class="onoffswitch2"><input type="checkbox" name="checkbox" id="myonoffswitch-1" class="onoffswitch2-checkbox">
+											<label for="myonoffswitch-1" class="onoffswitch2-label"></label>
+										</div>
+									</div>
+								</div>
+							</div>
+							<div class="skin-theme-switcher">
+								<div class="swichermainleft">
+									<h4>Leftmenu Styles Mode</h4>
+									<div class="switch_section">
+										<div class="switch-toggle d-flex">
+											<span class="mr-auto">Light Menu</span>
+											<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch1" class="onoffswitch2-checkbox">
+												<label for="myonoffswitch1" class="onoffswitch2-label"></label>
+											</div>
+										</div>
+										<div class="switch-toggle d-flex">
+											<span class="mr-auto">Dark Menu</span>
+											<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch2" class="onoffswitch2-checkbox">
+												<label for="myonoffswitch2" class="onoffswitch2-label"></label>
+											</div>
+										</div>
+										<div class="switch-toggle d-flex">
+											<span class="mr-auto">Color Menu</span>
+											<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch" class="onoffswitch2-checkbox">
+												<label for="myonoffswitch" class="onoffswitch2-label"></label>
+											</div>
+										</div>
+										<div class="switch-toggle d-flex">
+											<span class="mr-auto">Gradient-Color Menu</span>
+											<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch5" class="onoffswitch2-checkbox">
+												<label for="myonoffswitch5" class="onoffswitch2-label"></label>
+											</div>
+										</div>
+									</div>
+								</div>
+							</div>
+							<div class="swichermainleft">
+								<h4>Color-Skins</h4>
+								<div class="skin-body">
+									<a class="wscolorcode1 blackborder color1" href="#" data-theme="assets/colors/color1.css"></a>
+									<a class="wscolorcode1 blackborder color2" href="#" data-theme="assets/colors/color2.css"></a>
+									<a class="wscolorcode1 blackborder color3" href="#" data-theme="assets/colors/color3.css"></a>
+									<a class="wscolorcode1 blackborder color4" href="#" data-theme="assets/colors/color4.css"></a>
+									<a class="wscolorcode1 blackborder color5" href="#" data-theme="assets/colors/color5.css"></a>
+								</div>
+							</div>
+						</div>
+					</div>
+				</div>
+			</div>
+		</div>
+		<!-- End Switcher -->
+		
+        <!-- GLOBAL-LOADER -->
+        <div id="global-loader">
+            <img src="assets/images/loader.svg" class="loader-img" alt="Loader">    
+        </div>
+        <!-- /GLOBAL-LOADER -->
+        <!-- PAGE -->
+         <div class="page">
+         <div class="page-main">
+			
+            <!--APP-SIDEBAR-->
+                <div class="app-sidebar__overlay" data-toggle="sidebar"></div>
+                <aside class="app-sidebar">
+                    <div class="side-header">
+                        <a class="header-brand1" href="index.html">
+                            <img src="assets/images/brand/logo.png" class="header-brand-img desktop-logo" alt="logo">
+                            <img src="assets/images/brand/logo-1.png"  class="header-brand-img toggle-logo" alt="logo">
+                            <img src="assets/images/brand/logo-2.png" class="header-brand-img light-logo" alt="logo">
+                            <img src="assets/images/brand/logo-3.png" class="header-brand-img light-logo1" alt="logo">
+                        </a><!-- LOGO -->
+                        <a aria-label="Hide Sidebar" class="app-sidebar__toggle ml-auto" data-toggle="sidebar" href="#"></a><!-- sidebar-toggle-->
+					</div>
+                    <div class="app-sidebar__user">
+                        <div class="dropdown user-pro-body text-center">
+                            <div class="user-pic">
+                                <img src="assets/images/users/10.jpg" alt="user-img" class="avatar-xl rounded-circle">
+                            </div>
+                            <div class="user-info">
+                                <h6 class=" mb-0 text-dark">{{ Auth::user()->name }}</h6>
+                                <span class="text-muted app-sidebar__user-name text-sm">Administrator</span>
+                            </div>
+                        </div>
+					</div>
+					
+                    <div class="sidebar-navs">
+                        <ul class="nav  nav-pills-circle">
+                            <li class="nav-item" data-toggle="tooltip" data-placement="top" title="Settings">
+                                <a class="nav-link text-center m-2">
+                                    <i class="fe fe-settings"></i>
+                                </a>
+                            </li>
+                            <li class="nav-item" data-toggle="tooltip" data-placement="top" title="Chat">
+                                <a class="nav-link text-center m-2">
+                                    <i class="fe fe-mail"></i>
+                                </a>
+                            </li>
+                            <li class="nav-item" data-toggle="tooltip" data-placement="top" title="Followers">
+                                <a class="nav-link text-center m-2">
+                                    <i class="fe fe-user"></i>
+                                </a>
+                            </li>
+                            <li class="nav-item" data-toggle="tooltip" data-placement="top" title="Logout">
+								<a class="nav-link text-center m-2" href="{{ route('logout') }}" 
+								onclick="event.preventDefault(); document.getElementById('logout-form').submit();">
+                                    <i class="fe fe-power"></i>
+								</a>
+								<form id="logout-form" action="{{ route('logout') }}" method="POST" class="d-none">
+                                @csrf
+                            </form>
+                            </li>
+                        </ul>
+					</div>
+                    <ul class="side-menu">
+                        <li><h3>Main</h3></li>
+                        <li class="slide">
+							<a class="side-menu__item"  data-toggle="slide" href="{{ route('home') }}"><i class="side-menu__icon ti-home"></i><span class="side-menu__label">Dashboard</span></a>
+							<ul class="slide-menu">
+                                <li><a class="slide-item" href="{{ route('home') }}"><span>Sales Dashboard</span></a></li>
+                            </ul>
+                        </li>
+                        <li><h3>Inventory Users</h3></li>
+                        <li class="slide">
+                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-panel"></i><span class="side-menu__label">Users</span><i class="angle fa fa-angle-right"></i></a>
+                            <ul class="slide-menu">
+                                <li><a href="timeline.html" class="slide-item"> User Lists</a></li>
+                            </ul>
+						</li>
+						<li><h3>Inventory Purchase</h3></li>
+                        <li class="slide">
+                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-layers"></i><span class="side-menu__label">Sell</span><i class="angle fa fa-angle-right"></i></a>
+                            <ul class="slide-menu">
+                                <li><a href="alerts.html" class="slide-item"> New Sell</a></li>
+                                <li><a href="buttons.html" class="slide-item"> Inventory Sell Lists</a></li>
+                                <li><a href="colors.html" class="slide-item"> Invoice Lists</a></li>
+                                <li><a href="avatarsquare.html" class="slide-item"> Report</a></li>
+                            </ul>
+                        </li>
+                        <li class="slide">
+                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-rocket"></i><span class="side-menu__label">Inventory Order & Purchase</span><i class="angle fa fa-angle-right"></i></a>
+                            <ul class="slide-menu">
+                                <li><a href="mediaobject.html" class="slide-item"> New Orders</a></li>
+                                <li><a href="accordion.html" class="slide-item"> Orders List</a></li>
+                                <li><a href="tabs.html" class="slide-item"> Purchase List</a></li>
+                            </ul>
+                        </li>
+                        <li><h3>Supplier</h3></li>
+                        <li class="slide">
+                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-bar-chart"></i><span class="side-menu__label">Supplier</span><i class="angle fa fa-angle-right"></i></a>
+                            <ul class="slide-menu">
+                                <li><a href="chart-chartist.html" class="slide-item">Supplier Lists</a></li>
+                            </ul>
+                        </li>
+                        <li><h3>Customers</h3></li>
+                        <li class="slide">
+                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-receipt"></i><span class="side-menu__label">Customers</span><i class="angle fa fa-angle-right"></i></a>
+                            <ul class="slide-menu">
+                                <li><a href="form-elements.html" class="slide-item"> Ceramic Customers</a></li>
+                                <li><a href="form.html" class="slide-item"> Tiles Customres</a></li>
+                            </ul>
+                        </li>
+                        <li><h3>Employee Details</h3></li>
+                        <li class="slide">
+                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-receipt"></i><span class="side-menu__label">Employee Information</span><i class="angle fa fa-angle-right"></i></a>
+                            <ul class="slide-menu">
+                                <li><a href="form-elements.html" class="slide-item"> Employee Salery</a></li>
+                                <li><a href="form-elements.html" class="slide-item"> Utility</a></li>
+                                <li><a href="form.html" class="slide-item"> Expense Details</a></li>
+                            </ul>
+                        </li>
+                        <li><h3>Account Details</h3></li>
+                        <li class="slide">
+                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-receipt"></i><span class="side-menu__label">Account Information</span><i class="angle fa fa-angle-right"></i></a>
+                            <ul class="slide-menu">
+                                <li><a href="form-elements.html" class="slide-item"> Account Check</a></li>
+                                <li><a href="form-elements.html" class="slide-item"> Cash Details</a></li>
+                                <li><a href="form.html" class="slide-item"> Due Details</a></li>
+                                <li><a href="form.html" class="slide-item"> Customers Payment Details</a></li>
+                            </ul>
+                        </li>
+				
+<!--/APP-SIDEBAR-->
+
+            <!-- Mobile Header -->
+                <div class="mobile-header">
+                    <div class="container-fluid">
+                        <div class="d-flex">
+                            <a aria-label="Hide Sidebar" class="app-sidebar__toggle" data-toggle="sidebar" href="#"></a><!-- sidebar-toggle-->
+                            <a class="header-brand" href="index.html">
+                                <img src="assets/images/brand/logo.png" class="header-brand-img desktop-logo" alt="logo">
+                                <img src="assets/images/brand/logo-3.png" class="header-brand-img desktop-logo mobile-light" alt="logo">
+                            </a>
+                            <div class="d-flex order-lg-2 ml-auto header-right-icons">
+                                <button class="navbar-toggler navresponsive-toggler d-md-none" type="button" data-toggle="collapse" data-target="#navbarSupportedContent-4"
+                                    aria-controls="navbarSupportedContent-4" aria-expanded="false" aria-label="Toggle navigation">
+                                    <span class="navbar-toggler-icon fe fe-more-vertical text-white"></span>
+                                </button>
+                                <div class="dropdown profile-1">
+                                    <a href="#" data-toggle="dropdown" class="nav-link pr-2 leading-none d-flex">
+                                        <span>
+                                            <img src="assets/images/users/10.jpg" alt="profile-user" class="avatar  profile-user brround cover-image">
+                                        </span>
+                                    </a>
+                                    <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
+                                        <div class="drop-heading">
+                                            <div class="text-center">
+                                                <h5 class="text-dark mb-0">{{ Auth::user()->name }}</h5>
+                                                <small class="text-muted">Administrator</small>
+                                            </div>
+                                        </div>
+                                        <div class="dropdown-divider m-0"></div>
+                                        <a class="dropdown-item" href="#">
+                                            <i class="dropdown-icon mdi mdi-account-outline"></i> Profile
+                                        </a>
+                                        <a class="dropdown-item" href="#">
+                                            <i class="dropdown-icon  mdi mdi-settings"></i> Settings
+                                        </a>
+                                        <a class="dropdown-item" href="#">
+                                            <span class="float-right"></span>
+                                            <i class="dropdown-icon mdi  mdi-message-outline"></i> Inbox
+                                        </a>
+                                        <a class="dropdown-item" href="#">
+                                            <i class="dropdown-icon mdi mdi-comment-check-outline"></i> Message
+                                        </a>
+                                        <div class="dropdown-divider"></div>
+                                        <a class="dropdown-item" href="#">
+                                            <i class="dropdown-icon mdi mdi-compass-outline"></i> Need help?
+                                        </a>
+                                        <a class="dropdown-item" href="login.html">
+                                            <i class="dropdown-icon mdi  mdi-logout-variant"></i> Sign out
+                                        </a>
+                                    </div>
+                                </div>
+                                <div class="dropdown d-md-flex header-settings">
+                                    <a href="#" class="nav-link icon " data-toggle="sidebar-right" data-target=".sidebar-right">
+                                        <i class="fe fe-align-right"></i>
+                                    </a>
+                                </div><!-- SIDE-MENU -->
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                <div class="mb-1 navbar navbar-expand-lg  responsive-navbar navbar-dark d-md-none bg-white">
+                    <div class="collapse navbar-collapse" id="navbarSupportedContent-4">
+                        <div class="d-flex order-lg-2 ml-auto">
+                            <div class="dropdown d-sm-flex">
+                                <a href="#" class="nav-link icon" data-toggle="dropdown">
+                                    <i class="fe fe-search"></i>
+                                </a>
+                                <div class="dropdown-menu header-search dropdown-menu-left">
+                                    <div class="input-group w-100 p-2">
+                                        <input type="text" class="form-control " placeholder="Search....">
+                                        <div class="input-group-append ">
+                                            <button type="button" class="btn btn-primary ">
+                                                <i class="fa fa-search" aria-hidden="true"></i>
+                                            </button>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div><!-- SEARCH -->
+                            <div class="dropdown d-md-flex">
+                                <a class="nav-link icon full-screen-link nav-link-bg">
+                                    <i class="fe fe-maximize fullscreen-button"></i>
+                                </a>
+                            </div><!-- FULL-SCREEN -->
+                            <div class="dropdown d-md-flex notifications">
+                                <a class="nav-link icon" data-toggle="dropdown">
+                                    <i class="fe fe-bell"></i>
+                                </a>
+                                <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
+                                    <div class="notifications-menu">
+                                        <a class="dropdown-item d-flex pb-3" href="#">
+                                            <div class="fs-16 text-success mr-3">
+                                                <i class="fa fa-thumbs-o-up"></i>
+                                            </div>
+                                            <div class="">
+                                                <strong>Someone likes our posts.</strong>
+                                            </div>
+                                        </a>
+                                        <a class="dropdown-item d-flex pb-3" href="#">
+                                            <div class="fs-16 text-primary mr-3">
+                                                <i class="fa fa-commenting-o"></i>
+                                            </div>
+                                            <div class="">
+                                                <strong>3 New Comments.</strong>
+                                            </div>
+                                        </a>
+                                        <a class="dropdown-item d-flex pb-3" href="#">
+                                            <div class="fs-16 text-danger mr-3">
+                                                <i class="fa fa-cogs"></i>
+                                            </div>
+                                            <div class="">
+                                                <strong>Server Rebooted</strong>
+                                            </div>
+                                        </a>
+                                    </div>
+                                    <div class="dropdown-divider"></div>
+                                    <a href="#" class="dropdown-item text-center">View all Notification</a>
+                                </div>
+                            </div><!-- NOTIFICATIONS -->
+                            <div class="dropdown d-md-flex message">
+                                <a class="nav-link icon text-center" data-toggle="dropdown">
+                                    <i class="fe fe-mail"></i>
+                                </a>
+                                <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
+                                    <div class="message-menu">
+                                        <a class="dropdown-item d-flex pb-3" href="#">
+                                            <span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/1.jpg"></span>
+                                            <div>
+                                                <strong>Madeleine</strong> Hey! there I' am available....
+                                                <div class="small text-muted">
+                                                    3 hours ago
+                                                </div>
+                                            </div>
+                                        </a>
+                                        <a class="dropdown-item d-flex pb-3" href="#">
+                                            <span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/12.jpg"></span>
+                                            <div>
+                                                <strong>Anthony</strong> New product Launching...
+                                                <div class="small text-muted">
+                                                    5 hour ago
+                                                </div>
+                                            </div>
+                                        </a>
+                                        <a class="dropdown-item d-flex pb-3" href="#">
+                                            <span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/4.jpg"></span>
+                                            <div>
+                                                <strong>Olivia</strong> New Schedule Realease......
+                                                <div class="small text-muted">
+                                                    45 mintues ago
+                                                </div>
+                                            </div>
+                                        </a>
+                                        <a class="dropdown-item d-flex pb-3" href="#">
+                                            <span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/15.jpg"></span>
+                                            <div>
+                                                <strong>Sanderson</strong> New Schedule Realease......
+                                                <div class="small text-muted">
+                                                    2 days ago
+                                                </div>
+                                            </div>
+                                        </a>
+                                    </div>
+                                    <div class="dropdown-divider"></div>
+                                    <a href="#" class="dropdown-item text-center">See all Messages</a>
+                                </div>
+                            </div><!-- MESSAGE-BOX -->
+                        </div>
+                    </div>
+				</div>
+
+<!-- Header Components -->
