+++ resources/views/main/components/scripts.blade.php
@@ -1,0 +1,60 @@
+<!-- BACK-TO-TOP -->
+		<a href="#top" id="back-to-top"><i class="fa fa-angle-up"></i></a>
+		<!-- JQUERY JS -->
+		<script src="{{ asset('assets/js/jquery-3.4.1.min.js') }}"></script>
+
+		<!-- BOOTSTRAP JS -->
+		<script src="{{ asset('assets/plugins/bootstrap/js/bootstrap.bundle.min.js') }}"></script>
+		<script src="{{ asset('assets/plugins/bootstrap/js/popper.min.js') }}"></script>
+
+		<!-- SPARKLINE JS-->
+		<script src="{{ asset('assets/js/jquery.sparkline.min.js') }}"></script>
+
+		<!-- CHART-CIRCLE JS -->
+		<script src="{{ asset('assets/js/circle-progress.min.js') }}"></script>
+
+		<!-- RATING STAR JS-->
+		<script src="{{ asset('assets/plugins/rating/jquery.rating-stars.js') }}"></script>
+
+		<!-- C3.JS CHART JS -->
+		<script src="{{ asset('assets/plugins/charts-c3/d3.v5.min.js') }}"></script>
+		<script src="{{ asset('assets/plugins/charts-c3/c3-chart.js') }}"></script>
+
+		<!-- INPUT MASK JS-->
+		<script src="{{ asset('assets/plugins/input-mask/jquery.mask.min.js') }}"></script>
+
+		<!-- SIDE-MENU JS-->
+		<script src="{{ asset('assets/plugins/sidemenu/sidemenu.js') }}"></script>
+
+        <!-- CUSTOM SCROLLBAR JS-->
+		<script src="{{ asset('assets/plugins/scroll-bar/jquery.mCustomScrollbar.concat.min.js') }}"></script>
+		
+		<script src="{{ asset('assets/js/index1.js') }}"></script>
+<script src="{{ asset('assets/plugins/chart/Chart.bundle.js') }}"></script>
+<script src="{{ asset('assets/plugins/chart/utils.js') }}"></script>
+<script src="{{ asset('assets/js/apexcharts.js') }}"></script>
+<script src="{{ asset('assets/plugins/peitychart/jquery.peity.min.js') }}"></script>
+<script src="{{ asset('assets/plugins/peitychart/peitychart.init.js') }}"></script>
+<script src="{{asset('assets/plugins/echarts/echarts.js') }}"></script>
+
+		<!-- SIDEBAR JS -->
+		<script src="{{ asset('assets/plugins/sidebar/sidebar.js') }}"></script>
+		
+		<!-- P-scroll bar js-->
+		<script src="{{ asset('assets/plugins/p-scroll/perfect-scrollbar.js') }}"></script>
+		<script src="{{ asset('assets/plugins/p-scroll/p-scroll.js') }}"></script>
+		
+		<!-- SWITCHER JS -->
+		<script src="{{ asset('assets/switcher/js/switcher.js') }}"></script>
+		
+		<!--CUSTOM JS -->
+		<script src="{{ asset('assets/js/custom.js') }}"></script>
+
+		@stack('js')
+
+
+
+  
+    
+
+<!-- Mirrored from laravel.spruko.com/volgh/leftmenu_icon_light/index by HTTrack Website Copier/3.x [XR&CO'2014], Sat, 19 Dec 2020 16:20:50 GMT -->
