+++ resources/views/main/components/styles.blade.php
@@ -1,0 +1,49 @@
+<!-- META DATA -->
+        <meta charset="UTF-8">
+        <meta name='viewport' content='width=device-width, initial-scale=1.0, user-scalable=0'>
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <meta name="description" content="Volgh –  Bootstrap 4 Responsive Application Admin panel Theme Ui Kit & Premium Dashboard Design Modern Flat HTML Template">
+        <meta name="author" content="Spruko Technologies Private Limited">
+        <meta name="keywords" content="analytics dashboard, bootstrap 4 web app admin template, bootstrap admin panel, bootstrap admin template, bootstrap dashboard, bootstrap panel, Application dashboard design, dashboard design template, dashboard jquery clean html, dashboard template theme, dashboard responsive ui, html admin backend template ui kit, html flat dashboard template, it admin dashboard ui, premium modern html template">
+        <!-- FAVICON -->
+        <link rel="shortcut icon" type="image/x-icon" href="{{ asset('assets/images/brand/favicon.ico') }}" />
+
+        <!-- TITLE -->
+        <title>Inventory Management</title>
+
+        
+        <!-- BOOTSTRAP CSS -->
+        <link href="{{ asset('assets/plugins/bootstrap/css/bootstrap.min.css') }}" rel="stylesheet" />
+
+        <!-- STYLE CSS -->
+        <link href="{{ asset('assets/css/style.css') }}" rel="stylesheet"/>
+        <link href="{{ asset('assets/css-dark/dark-style.css') }}" rel="stylesheet"/>
+        <link href="{{ asset('assets/css/skin-modes.css') }}" rel="stylesheet"/>
+		
+		<!-- P-scroll bar css-->
+		<link href="{{ asset('assets/plugins/p-scroll/perfect-scrollbar.css') }}" rel="stylesheet" />
+
+        <!-- SIDE-MENU CSS -->
+        <link href="{{ asset('assets/plugins/sidemenu/sidemenu.css') }}" rel="stylesheet">
+
+        <!--C3.JS CHARTS PLUGIN -->
+        <link href="{{ asset('assets/plugins/charts-c3/c3-chart.css') }}" rel="stylesheet"/>
+        
+        
+        <!-- CUSTOM SCROLL BAR CSS-->
+        <link href="{{ asset('assets/plugins/scroll-bar/jquery.mCustomScrollbar.css') }}" rel="stylesheet"/>
+
+        <!--- FONT-ICONS CSS -->
+        <link href="{{ asset('assets/css/icons.css') }}" rel="stylesheet"/>
+
+        <!-- SIDEBAR CSS -->
+        <link href="{{ asset('assets/plugins/sidebar/sidebar.css') }}" rel="stylesheet">
+
+        <!-- COLOR SKIN CSS -->
+        <link id="theme" rel="stylesheet" type="text/css" media="all" href="{{ asset('assets/colors/color1.css') }}" />
+
+		<!-- SWITCHER  CSS -->
+        <link  href="{{ asset('assets/switcher/css/switcher.css') }}" rel="stylesheet"/>
+		<link  href="{{ asset('assets/switcher/demo.css') }}" rel="stylesheet"/>      
+		
+		@stack('css')
