--- resources/views/main/dashboard.blade.php
@@ -1,1088 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-    
-<!-- Mirrored from laravel.spruko.com/volgh/leftmenu_icon_light/index by HTTrack Website Copier/3.x [XR&CO'2014], Sat, 19 Dec 2020 16:19:56 GMT -->
-<!-- Added by HTTrack --><meta http-equiv="content-type" content="text/html;charset=UTF-8" /><!-- /Added by HTTrack -->
-<head>
-       
-        <!-- META DATA -->
-        <meta charset="UTF-8">
-        <meta name='viewport' content='width=device-width, initial-scale=1.0, user-scalable=0'>
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="description" content="Volgh –  Bootstrap 4 Responsive Application Admin panel Theme Ui Kit & Premium Dashboard Design Modern Flat HTML Template">
-        <meta name="author" content="Spruko Technologies Private Limited">
-        <meta name="keywords" content="analytics dashboard, bootstrap 4 web app admin template, bootstrap admin panel, bootstrap admin template, bootstrap dashboard, bootstrap panel, Application dashboard design, dashboard design template, dashboard jquery clean html, dashboard template theme, dashboard responsive ui, html admin backend template ui kit, html flat dashboard template, it admin dashboard ui, premium modern html template">
-        <!-- FAVICON -->
-        <link rel="shortcut icon" type="image/x-icon" href="{{ asset('assets/images/brand/favicon.ico') }}" />
-
-        <!-- TITLE -->
-        <title>Inventory Management</title>
-
-        
-        <!-- BOOTSTRAP CSS -->
-        <link href="{{ asset('assets/plugins/bootstrap/css/bootstrap.min.css') }}" rel="stylesheet" />
-
-        <!-- STYLE CSS -->
-        <link href="{{ asset('assets/css/style.css') }}" rel="stylesheet"/>
-        <link href="{{ asset('assets/css-dark/dark-style.css') }}" rel="stylesheet"/>
-        <link href="{{ asset('assets/css/skin-modes.css') }}" rel="stylesheet"/>
-		
-		<!-- P-scroll bar css-->
-		<link href="{{ asset('assets/plugins/p-scroll/perfect-scrollbar.css') }}" rel="stylesheet" />
-
-        <!-- SIDE-MENU CSS -->
-        <link href="{{ asset('assets/plugins/sidemenu/sidemenu.css') }}" rel="stylesheet">
-
-        <!--C3.JS CHARTS PLUGIN -->
-        <link href="{{ asset('assets/plugins/charts-c3/c3-chart.css') }}" rel="stylesheet"/>
-        
-        
-        <!-- CUSTOM SCROLL BAR CSS-->
-        <link href="{{ asset('assets/plugins/scroll-bar/jquery.mCustomScrollbar.css') }}" rel="stylesheet"/>
-
-        <!--- FONT-ICONS CSS -->
-        <link href="{{ asset('assets/css/icons.css') }}" rel="stylesheet"/>
-
-        <!-- SIDEBAR CSS -->
-        <link href="{{ asset('assets/plugins/sidebar/sidebar.css') }}" rel="stylesheet">
-
-        <!-- COLOR SKIN CSS -->
-        <link id="theme" rel="stylesheet" type="text/css" media="all" href="{{ asset('assets/colors/color1.css') }}" />
-
-		<!-- SWITCHER  CSS -->
-        <link  href="{{ asset('assets/switcher/css/switcher.css') }}" rel="stylesheet"/>
-        <link  href="{{ asset('assets/switcher/demo.css') }}" rel="stylesheet"/>       
-    </head>
-
-    <body class="app sidebar-mini">
-		
-		<!-- Start Switcher -->
-		<div class="switcher-wrapper ">
-			<div class="demo_changer">
-				<div class="demo-icon bg_dark"><i class="fa fa-cog fa-spin  text_primary"></i></div>
-				<div class="form_holder sidebar-right1">
-					<div class="row">
-						<div class="predefined_styles">
-							<div class="swichermainleft border-bottom text-center">
-								<div class="p-3">
-									<a href="https://codecanyon.net/user/sprukosoft/portfolio" class="btn btn-danger btn-block mt-0">View Demo</a>
-									<a href="https://codecanyon.net/item/volgh-laravel-admin-template/25860093" class="btn btn-success btn-block">Buy Now</a>
-									<a href="https://themeforest.net/user/sprukosoft/portfolio" class="btn btn-info btn-block">Our Portfolio</a>
-								</div>
-							</div>
-							<div class="swichermainleft">
-								<h4>Navigation Style</h4>
-								<div class="pl-3 pr-3">
-									<a class="btn btn-primary btn-block" href="https://laravel.spruko.com/volgh/horizontal_light/index">
-										Horizontal
-									</a>
-									<a class="btn btn-secondary btn-block" href="index.html">
-										Left-menu
-									</a>
-								</div>
-							</div>
-							<div class="swichermainleft">
-								<h4>Skin Modes</h4>
-								<div class="switch_section">
-									<div class="switch-toggle d-flex">
-										<span class="mr-auto">Light Mode</span>
-										<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch3" class="onoffswitch2-checkbox" checked>
-											<label for="myonoffswitch3" class="onoffswitch2-label"></label>
-										</div>
-									</div>
-									<div class="switch-toggle d-flex">
-										<span class="mr-auto">Transparent Mode</span>
-										<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch4" class="onoffswitch2-checkbox">
-											<label for="myonoffswitch4" class="onoffswitch2-label"></label>
-										</div>
-									</div>
-									<div class="switch-toggle d-flex">
-										<span class="mr-auto">Dark Mode</span>
-										<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch16" class="onoffswitch2-checkbox">
-											<label for="myonoffswitch16" class="onoffswitch2-label"></label>
-										</div>
-									</div>
-								</div>
-							</div>
-							<div class="clearfix"></div>
-							<div class="swichermainleft">
-								<div class="switch_bg-img">
-									<div class=" d-flex">
-										<span class="mr-auto">Left-menu Bg-Image</span>
-										<div class="onoffswitch2"><input type="checkbox" name="checkbox" id="myonoffswitch-1" class="onoffswitch2-checkbox">
-											<label for="myonoffswitch-1" class="onoffswitch2-label"></label>
-										</div>
-									</div>
-								</div>
-							</div>
-							<div class="skin-theme-switcher">
-								<div class="swichermainleft">
-									<h4>Leftmenu Styles Mode</h4>
-									<div class="switch_section">
-										<div class="switch-toggle d-flex">
-											<span class="mr-auto">Light Menu</span>
-											<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch1" class="onoffswitch2-checkbox">
-												<label for="myonoffswitch1" class="onoffswitch2-label"></label>
-											</div>
-										</div>
-										<div class="switch-toggle d-flex">
-											<span class="mr-auto">Dark Menu</span>
-											<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch2" class="onoffswitch2-checkbox">
-												<label for="myonoffswitch2" class="onoffswitch2-label"></label>
-											</div>
-										</div>
-										<div class="switch-toggle d-flex">
-											<span class="mr-auto">Color Menu</span>
-											<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch" class="onoffswitch2-checkbox">
-												<label for="myonoffswitch" class="onoffswitch2-label"></label>
-											</div>
-										</div>
-										<div class="switch-toggle d-flex">
-											<span class="mr-auto">Gradient-Color Menu</span>
-											<div class="onoffswitch2"><input type="radio" name="onoffswitch2" id="myonoffswitch5" class="onoffswitch2-checkbox">
-												<label for="myonoffswitch5" class="onoffswitch2-label"></label>
-											</div>
-										</div>
-									</div>
-								</div>
-							</div>
-							<div class="swichermainleft">
-								<h4>Color-Skins</h4>
-								<div class="skin-body">
-									<a class="wscolorcode1 blackborder color1" href="#" data-theme="assets/colors/color1.css"></a>
-									<a class="wscolorcode1 blackborder color2" href="#" data-theme="assets/colors/color2.css"></a>
-									<a class="wscolorcode1 blackborder color3" href="#" data-theme="assets/colors/color3.css"></a>
-									<a class="wscolorcode1 blackborder color4" href="#" data-theme="assets/colors/color4.css"></a>
-									<a class="wscolorcode1 blackborder color5" href="#" data-theme="assets/colors/color5.css"></a>
-								</div>
-							</div>
-						</div>
-					</div>
-				</div>
-			</div>
-		</div>
-		<!-- End Switcher -->
-		
-        <!-- GLOBAL-LOADER -->
-        <div id="global-loader">
-            <img src="assets/images/loader.svg" class="loader-img" alt="Loader">    
-        </div>
-        <!-- /GLOBAL-LOADER -->
-        <!-- PAGE -->
-         <div class="page">
-         <div class="page-main">
-			
-            <!--APP-SIDEBAR-->
-                <div class="app-sidebar__overlay" data-toggle="sidebar"></div>
-                <aside class="app-sidebar">
-                    <div class="side-header">
-                        <a class="header-brand1" href="index.html">
-                            <img src="assets/images/brand/logo.png" class="header-brand-img desktop-logo" alt="logo">
-                            <img src="assets/images/brand/logo-1.png"  class="header-brand-img toggle-logo" alt="logo">
-                            <img src="assets/images/brand/logo-2.png" class="header-brand-img light-logo" alt="logo">
-                            <img src="assets/images/brand/logo-3.png" class="header-brand-img light-logo1" alt="logo">
-                        </a><!-- LOGO -->
-                        <a aria-label="Hide Sidebar" class="app-sidebar__toggle ml-auto" data-toggle="sidebar" href="#"></a><!-- sidebar-toggle-->
-					</div>
-                    <div class="app-sidebar__user">
-                        <div class="dropdown user-pro-body text-center">
-                            <div class="user-pic">
-                                <img src="assets/images/users/10.jpg" alt="user-img" class="avatar-xl rounded-circle">
-                            </div>
-                            <div class="user-info">
-                                <h6 class=" mb-0 text-dark">{{ Auth::user()->name }}</h6>
-                                <span class="text-muted app-sidebar__user-name text-sm">Administrator</span>
-                            </div>
-                        </div>
-					</div>
-					
-                    <div class="sidebar-navs">
-                        <ul class="nav  nav-pills-circle">
-                            <li class="nav-item" data-toggle="tooltip" data-placement="top" title="Settings">
-                                <a class="nav-link text-center m-2">
-                                    <i class="fe fe-settings"></i>
-                                </a>
-                            </li>
-                            <li class="nav-item" data-toggle="tooltip" data-placement="top" title="Chat">
-                                <a class="nav-link text-center m-2">
-                                    <i class="fe fe-mail"></i>
-                                </a>
-                            </li>
-                            <li class="nav-item" data-toggle="tooltip" data-placement="top" title="Followers">
-                                <a class="nav-link text-center m-2">
-                                    <i class="fe fe-user"></i>
-                                </a>
-                            </li>
-                            <li class="nav-item" data-toggle="tooltip" data-placement="top" title="Logout">
-								<a class="nav-link text-center m-2" href="{{ route('logout') }}" 
-								onclick="event.preventDefault(); document.getElementById('logout-form').submit();">
-                                    <i class="fe fe-power"></i>
-								</a>
-								<form id="logout-form" action="{{ route('logout') }}" method="POST" class="d-none">
-                                @csrf
-                            </form>
-                            </li>
-                        </ul>
-					</div>
-                    <ul class="side-menu">
-                        <li><h3>Main</h3></li>
-                        <li class="slide">
-							<a class="side-menu__item"  data-toggle="slide" href="#"><i class="side-menu__icon ti-home"></i><span class="side-menu__label">Dashboard</span></a>
-							<ul class="slide-menu">
-                                <li><a class="slide-item" href="index.html"><span>Sales Dashboard</span></a></li>
-                            </ul>
-                        </li>
-                        <li><h3>Inventory Users</h3></li>
-                        <li class="slide">
-                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-panel"></i><span class="side-menu__label">Users</span><i class="angle fa fa-angle-right"></i></a>
-                            <ul class="slide-menu">
-                                <li><a href="timeline.html" class="slide-item"> User Lists</a></li>
-                            </ul>
-						</li>
-						<li><h3>Inventory Purchase</h3></li>
-                        <li class="slide">
-                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-layers"></i><span class="side-menu__label">Sell</span><i class="angle fa fa-angle-right"></i></a>
-                            <ul class="slide-menu">
-                                <li><a href="alerts.html" class="slide-item"> New Sell</a></li>
-                                <li><a href="buttons.html" class="slide-item"> Inventory Sell Lists</a></li>
-                                <li><a href="colors.html" class="slide-item"> Invoice Lists</a></li>
-                                <li><a href="avatarsquare.html" class="slide-item"> Report</a></li>
-                            </ul>
-                        </li>
-                        <li class="slide">
-                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-rocket"></i><span class="side-menu__label">Inventory Order & Purchase</span><i class="angle fa fa-angle-right"></i></a>
-                            <ul class="slide-menu">
-                                <li><a href="mediaobject.html" class="slide-item"> New Orders</a></li>
-                                <li><a href="accordion.html" class="slide-item"> Orders List</a></li>
-                                <li><a href="tabs.html" class="slide-item"> Purchase List</a></li>
-                            </ul>
-                        </li>
-                        <li><h3>Supplier</h3></li>
-                        <li class="slide">
-                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-bar-chart"></i><span class="side-menu__label">Supplier</span><i class="angle fa fa-angle-right"></i></a>
-                            <ul class="slide-menu">
-                                <li><a href="chart-chartist.html" class="slide-item">Supplier Lists</a></li>
-                            </ul>
-                        </li>
-                        <li><h3>Customers</h3></li>
-                        <li class="slide">
-                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-receipt"></i><span class="side-menu__label">Customers</span><i class="angle fa fa-angle-right"></i></a>
-                            <ul class="slide-menu">
-                                <li><a href="form-elements.html" class="slide-item"> Ceramic Customers</a></li>
-                                <li><a href="form.html" class="slide-item"> Tiles Customres</a></li>
-                            </ul>
-                        </li>
-                        <li><h3>Employee Details</h3></li>
-                        <li class="slide">
-                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-receipt"></i><span class="side-menu__label">Employee Information</span><i class="angle fa fa-angle-right"></i></a>
-                            <ul class="slide-menu">
-                                <li><a href="form-elements.html" class="slide-item"> Employee Salery</a></li>
-                                <li><a href="form-elements.html" class="slide-item"> Utility</a></li>
-                                <li><a href="form.html" class="slide-item"> Expense Details</a></li>
-                            </ul>
-                        </li>
-                        <li><h3>Account Details</h3></li>
-                        <li class="slide">
-                            <a class="side-menu__item" data-toggle="slide" href="#"><i class="side-menu__icon ti-receipt"></i><span class="side-menu__label">Account Information</span><i class="angle fa fa-angle-right"></i></a>
-                            <ul class="slide-menu">
-                                <li><a href="form-elements.html" class="slide-item"> Account Check</a></li>
-                                <li><a href="form-elements.html" class="slide-item"> Cash Details</a></li>
-                                <li><a href="form.html" class="slide-item"> Due Details</a></li>
-                                <li><a href="form.html" class="slide-item"> Customers Payment Details</a></li>
-                            </ul>
-                        </li>
-				
-<!--/APP-SIDEBAR-->
-            <!-- Mobile Header -->
-                <div class="mobile-header">
-                    <div class="container-fluid">
-                        <div class="d-flex">
-                            <a aria-label="Hide Sidebar" class="app-sidebar__toggle" data-toggle="sidebar" href="#"></a><!-- sidebar-toggle-->
-                            <a class="header-brand" href="index.html">
-                                <img src="assets/images/brand/logo.png" class="header-brand-img desktop-logo" alt="logo">
-                                <img src="assets/images/brand/logo-3.png" class="header-brand-img desktop-logo mobile-light" alt="logo">
-                            </a>
-                            <div class="d-flex order-lg-2 ml-auto header-right-icons">
-                                <button class="navbar-toggler navresponsive-toggler d-md-none" type="button" data-toggle="collapse" data-target="#navbarSupportedContent-4"
-                                    aria-controls="navbarSupportedContent-4" aria-expanded="false" aria-label="Toggle navigation">
-                                    <span class="navbar-toggler-icon fe fe-more-vertical text-white"></span>
-                                </button>
-                                <div class="dropdown profile-1">
-                                    <a href="#" data-toggle="dropdown" class="nav-link pr-2 leading-none d-flex">
-                                        <span>
-                                            <img src="assets/images/users/10.jpg" alt="profile-user" class="avatar  profile-user brround cover-image">
-                                        </span>
-                                    </a>
-                                    <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
-                                        <div class="drop-heading">
-                                            <div class="text-center">
-                                                <h5 class="text-dark mb-0">{{ Auth::user()->name }}</h5>
-                                                <small class="text-muted">Administrator</small>
-                                            </div>
-                                        </div>
-                                        <div class="dropdown-divider m-0"></div>
-                                        <a class="dropdown-item" href="#">
-                                            <i class="dropdown-icon mdi mdi-account-outline"></i> Profile
-                                        </a>
-                                        <a class="dropdown-item" href="#">
-                                            <i class="dropdown-icon  mdi mdi-settings"></i> Settings
-                                        </a>
-                                        <a class="dropdown-item" href="#">
-                                            <span class="float-right"></span>
-                                            <i class="dropdown-icon mdi  mdi-message-outline"></i> Inbox
-                                        </a>
-                                        <a class="dropdown-item" href="#">
-                                            <i class="dropdown-icon mdi mdi-comment-check-outline"></i> Message
-                                        </a>
-                                        <div class="dropdown-divider"></div>
-                                        <a class="dropdown-item" href="#">
-                                            <i class="dropdown-icon mdi mdi-compass-outline"></i> Need help?
-                                        </a>
-                                        <a class="dropdown-item" href="login.html">
-                                            <i class="dropdown-icon mdi  mdi-logout-variant"></i> Sign out
-                                        </a>
-                                    </div>
-                                </div>
-                                <div class="dropdown d-md-flex header-settings">
-                                    <a href="#" class="nav-link icon " data-toggle="sidebar-right" data-target=".sidebar-right">
-                                        <i class="fe fe-align-right"></i>
-                                    </a>
-                                </div><!-- SIDE-MENU -->
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="mb-1 navbar navbar-expand-lg  responsive-navbar navbar-dark d-md-none bg-white">
-                    <div class="collapse navbar-collapse" id="navbarSupportedContent-4">
-                        <div class="d-flex order-lg-2 ml-auto">
-                            <div class="dropdown d-sm-flex">
-                                <a href="#" class="nav-link icon" data-toggle="dropdown">
-                                    <i class="fe fe-search"></i>
-                                </a>
-                                <div class="dropdown-menu header-search dropdown-menu-left">
-                                    <div class="input-group w-100 p-2">
-                                        <input type="text" class="form-control " placeholder="Search....">
-                                        <div class="input-group-append ">
-                                            <button type="button" class="btn btn-primary ">
-                                                <i class="fa fa-search" aria-hidden="true"></i>
-                                            </button>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div><!-- SEARCH -->
-                            <div class="dropdown d-md-flex">
-                                <a class="nav-link icon full-screen-link nav-link-bg">
-                                    <i class="fe fe-maximize fullscreen-button"></i>
-                                </a>
-                            </div><!-- FULL-SCREEN -->
-                            <div class="dropdown d-md-flex notifications">
-                                <a class="nav-link icon" data-toggle="dropdown">
-                                    <i class="fe fe-bell"></i>
-                                </a>
-                                <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
-                                    <div class="notifications-menu">
-                                        <a class="dropdown-item d-flex pb-3" href="#">
-                                            <div class="fs-16 text-success mr-3">
-                                                <i class="fa fa-thumbs-o-up"></i>
-                                            </div>
-                                            <div class="">
-                                                <strong>Someone likes our posts.</strong>
-                                            </div>
-                                        </a>
-                                        <a class="dropdown-item d-flex pb-3" href="#">
-                                            <div class="fs-16 text-primary mr-3">
-                                                <i class="fa fa-commenting-o"></i>
-                                            </div>
-                                            <div class="">
-                                                <strong>3 New Comments.</strong>
-                                            </div>
-                                        </a>
-                                        <a class="dropdown-item d-flex pb-3" href="#">
-                                            <div class="fs-16 text-danger mr-3">
-                                                <i class="fa fa-cogs"></i>
-                                            </div>
-                                            <div class="">
-                                                <strong>Server Rebooted</strong>
-                                            </div>
-                                        </a>
-                                    </div>
-                                    <div class="dropdown-divider"></div>
-                                    <a href="#" class="dropdown-item text-center">View all Notification</a>
-                                </div>
-                            </div><!-- NOTIFICATIONS -->
-                            <div class="dropdown d-md-flex message">
-                                <a class="nav-link icon text-center" data-toggle="dropdown">
-                                    <i class="fe fe-mail"></i>
-                                </a>
-                                <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
-                                    <div class="message-menu">
-                                        <a class="dropdown-item d-flex pb-3" href="#">
-                                            <span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/1.jpg"></span>
-                                            <div>
-                                                <strong>Madeleine</strong> Hey! there I' am available....
-                                                <div class="small text-muted">
-                                                    3 hours ago
-                                                </div>
-                                            </div>
-                                        </a>
-                                        <a class="dropdown-item d-flex pb-3" href="#">
-                                            <span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/12.jpg"></span>
-                                            <div>
-                                                <strong>Anthony</strong> New product Launching...
-                                                <div class="small text-muted">
-                                                    5 hour ago
-                                                </div>
-                                            </div>
-                                        </a>
-                                        <a class="dropdown-item d-flex pb-3" href="#">
-                                            <span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/4.jpg"></span>
-                                            <div>
-                                                <strong>Olivia</strong> New Schedule Realease......
-                                                <div class="small text-muted">
-                                                    45 mintues ago
-                                                </div>
-                                            </div>
-                                        </a>
-                                        <a class="dropdown-item d-flex pb-3" href="#">
-                                            <span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/15.jpg"></span>
-                                            <div>
-                                                <strong>Sanderson</strong> New Schedule Realease......
-                                                <div class="small text-muted">
-                                                    2 days ago
-                                                </div>
-                                            </div>
-                                        </a>
-                                    </div>
-                                    <div class="dropdown-divider"></div>
-                                    <a href="#" class="dropdown-item text-center">See all Messages</a>
-                                </div>
-                            </div><!-- MESSAGE-BOX -->
-                        </div>
-                    </div>
-                </div>
-<!-- /Mobile Header -->
-        <div class="app-content">
-        <div class="side-app">
-        <div class="page-header">    
-                                    <!-- PAGE-HEADER -->
-							<div>
-								<h1 class="page-title">Dashboard 01</h1>
-								<ol class="breadcrumb">
-									<li class="breadcrumb-item"><a href="#">Home</a></li>
-									<li class="breadcrumb-item active" aria-current="page">Dashboard 01</li>
-								</ol>
-							</div>							
-						<!-- PAGE-HEADER END -->
-            <!-- /Notification -->
-							<div class="d-flex  ml-auto header-right-icons header-search-icon">
-								<div class="dropdown d-sm-flex">
-									<a href="#" class="nav-link icon" data-toggle="dropdown">
-										<i class="fe fe-search"></i>
-									</a>
-									<div class="dropdown-menu header-search dropdown-menu-left">
-										<div class="input-group w-100 p-2">
-											<input type="text" class="form-control " placeholder="Search....">
-											<div class="input-group-append ">
-												<button type="button" class="btn btn-primary ">
-													<i class="fa fa-search" aria-hidden="true"></i>
-												</button>
-											</div>
-										</div>
-									</div>
-								</div><!-- SEARCH -->
-								<div class="dropdown d-md-flex">
-									<a class="nav-link icon full-screen-link nav-link-bg">
-										<i class="fe fe-maximize fullscreen-button"></i>
-									</a>
-								</div><!-- FULL-SCREEN -->
-								<div class="dropdown d-md-flex notifications">
-									<a class="nav-link icon" data-toggle="dropdown">
-										<i class="fe fe-bell"></i>
-										<span class="nav-unread badge badge-success badge-pill">2</span>
-									</a>
-									<div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
-										<div class="notifications-menu">
-											<a class="dropdown-item d-flex pb-3" href="#">
-												<div class="fs-16 text-success mr-3">
-													<i class="fa fa-thumbs-o-up"></i>
-												</div>
-												<div class="">
-													<strong>Someone likes our posts.</strong>
-												</div>
-											</a>
-											<a class="dropdown-item d-flex pb-3" href="#">
-												<div class="fs-16 text-primary mr-3">
-													<i class="fa fa-commenting-o"></i>
-												</div>
-												<div class="">
-													<strong>3 New Comments.</strong>
-												</div>
-											</a>
-											<a class="dropdown-item d-flex pb-3" href="#">
-												<div class="fs-16 text-danger mr-3">
-													<i class="fa fa-cogs"></i>
-												</div>
-												<div class="">
-													<strong>Server Rebooted</strong>
-												</div>
-											</a>
-										</div>
-										<div class="dropdown-divider"></div>
-										<a href="#" class="dropdown-item text-center">View all Notification</a>
-									</div>
-								</div><!-- NOTIFICATIONS -->
-								<div class="dropdown d-md-flex message">
-									<a class="nav-link icon text-center" data-toggle="dropdown">
-										<i class="fe fe-mail"></i>
-										<span class="nav-unread badge badge-danger badge-pill">3</span>
-									</a>
-									<div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
-										<div class="message-menu">
-											<a class="dropdown-item d-flex pb-3" href="#">
-												<span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/1.jpg"></span>
-												<div>
-													<strong>Madeleine</strong> Hey! there I' am available....
-													<div class="small text-muted">
-														3 hours ago
-													</div>
-												</div>
-											</a>
-											<a class="dropdown-item d-flex pb-3" href="#">
-												<span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/12.jpg"></span>
-												<div>
-													<strong>Anthony</strong> New product Launching...
-													<div class="small text-muted">
-														5 hour ago
-													</div>
-												</div>
-											</a>
-											<a class="dropdown-item d-flex pb-3" href="#">
-												<span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/4.jpg"></span>
-												<div>
-													<strong>Olivia</strong> New Schedule Realease......
-													<div class="small text-muted">
-														45 mintues ago
-													</div>
-												</div>
-											</a>
-											<a class="dropdown-item d-flex pb-3" href="#">
-												<span class="avatar avatar-md brround mr-3 align-self-center cover-image" data-image-src="assets/images/users/15.jpg"></span>
-												<div>
-													<strong>Sanderson</strong> New Schedule Realease......
-													<div class="small text-muted">
-														2 days ago
-													</div>
-												</div>
-											</a>
-										</div>
-										<div class="dropdown-divider"></div>
-										<a href="#" class="dropdown-item text-center">See all Messages</a>
-									</div>
-								</div><!-- MESSAGE-BOX -->
-								<div class="dropdown profile-1">
-									<a href="#" data-toggle="dropdown" class="nav-link pr-2 leading-none d-flex">
-										<span>
-											<img src="assets/images/users/10.jpg" alt="profile-user" class="avatar  profile-user brround cover-image">
-										</span>
-									</a>
-									<div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow">
-										<div class="drop-heading">
-											<div class="text-center">
-												<h5 class="text-dark mb-0">Elizabeth Dyer</h5>
-												<small class="text-muted">Administrator</small>
-											</div>
-										</div>
-										<div class="dropdown-divider m-0"></div>
-										<a class="dropdown-item" href="#">
-											<i class="dropdown-icon mdi mdi-account-outline"></i> Profile
-										</a>
-										<a class="dropdown-item" href="#">
-											<i class="dropdown-icon  mdi mdi-settings"></i> Settings
-										</a>
-										<a class="dropdown-item" href="#">
-											<span class="float-right"></span>
-											<i class="dropdown-icon mdi  mdi-message-outline"></i> Inbox
-										</a>
-										<a class="dropdown-item" href="#">
-											<i class="dropdown-icon mdi mdi-comment-check-outline"></i> Message
-										</a>
-										<div class="dropdown-divider mt-0"></div>
-										<a class="dropdown-item" href="#">
-											<i class="dropdown-icon mdi mdi-compass-outline"></i> Need help?
-										</a>
-										<a class="dropdown-item" href="login.html">
-											<i class="dropdown-icon mdi  mdi-logout-variant"></i> Sign out
-										</a>
-									</div>
-								</div>
-								<div class="dropdown d-md-flex header-settings">
-									<a href="#" class="nav-link icon " data-toggle="sidebar-right" data-target=".sidebar-right">
-										<i class="fe fe-align-right"></i>
-									</a>
-								</div><!-- SIDE-MENU -->
-							</div>
-<!-- /Notification Ends -->
-  
-        </div> 
-            						<!-- ROW-1 -->
-						<div class="row">
-							<div class="col-lg-12 col-md-12 col-sm-12 col-xl-6">
-								<div class="row">
-									<div class="col-lg-6 col-md-12 col-sm-12 col-xl-6">
-										<div class="card">
-											<div class="card-body text-center statistics-info">
-												<div class="counter-icon bg-primary mb-0 box-primary-shadow">
-													<i class="fe fe-trending-up text-white"></i>
-												</div>
-												<h6 class="mt-4 mb-1">Total Sales</h6>
-												<h2 class="mb-2 number-font">34,516</h2>
-												<p class="text-muted">Sed ut perspiciatis unde omnis accusantium doloremque</p>
-											</div>
-										</div>
-									</div>
-									<div class="col-lg-6 col-md-12 col-sm-12 col-xl-6">
-										<div class="card">
-											<div class="card-body text-center statistics-info">
-												<div class="counter-icon bg-secondary mb-0 box-secondary-shadow">
-													<i class="fe fe-codepen text-white"></i>
-												</div>
-												<h6 class="mt-4 mb-1">Total Leads</h6>
-												<h2 class="mb-2 number-font">56,992</h2>
-												<p class="text-muted">Sed ut perspiciatis unde omnis accusantium doloremque</p>
-											</div>
-										</div>
-									</div>
-									<div class="col-lg-6 col-md-12 col-sm-12 col-xl-6">
-										<div class="card">
-											<div class="card-body text-center statistics-info">
-												<div class="counter-icon bg-success mb-0 box-success-shadow">
-													<i class="fe fe-dollar-sign text-white"></i>
-												</div>
-												<h6 class="mt-4 mb-1">Total Profit</h6>
-												<h2 class="mb-2  number-font">$42,567</h2>
-												<p class="text-muted">Sed ut perspiciatis unde omnis accusantium doloremque</p>
-											</div>
-										</div>
-									</div>
-									<div class="col-lg-6 col-md-12 col-sm-12 col-xl-6">
-										<div class="card">
-											<div class="card-body text-center statistics-info">
-												<div class="counter-icon bg-info mb-0 box-info-shadow">
-													<i class="fe fe-briefcase text-white"></i>
-												</div>
-												<h6 class="mt-4 mb-1">Total Cost</h6>
-												<h2 class="mb-2  number-font">$34,789</h2>
-												<p class="text-muted">Sed ut perspiciatis unde omnis accusantium doloremque</p>
-											</div>
-										</div>
-									</div>
-								</div>
-							</div>
-							<div class="col-sm-12 col-md-12 col-lg-12 col-xl-6">
-								<div class="card">
-									<div class="card-header">
-										<h3 class="card-title">Earnings</h3>
-									</div>
-									<div class="card-body">
-										<div id="echart1" class="chart-donut chart-dropshadow"></div>
-										<div class="mt-4">
-											<span class="ml-5"><span class="dot-label bg-info mr-2"></span>Sales</span>
-											<span class="ml-5"><span class="dot-label bg-secondary mr-2"></span>Profit</span>
-											<span class="ml-5"><span class="dot-label bg-success mr-2"></span>Growth</span>
-										</div>
-									</div>
-								</div>
-							</div><!-- COL END -->
-						</div>
-						<!-- ROW-1 END -->
-
-						<!-- ROW-3 -->
-						<div class="row">
-							<div class="col-sm-12 col-md-12 col-lg-12 col-xl-4">
-								<div class="card">
-									<div class="card-header">
-										<h3 class="card-title">Browsing Status</h3>
-									</div>
-									<div class="card-body p-4">
-										<div class="mb-5">
-											<p class="mb-2">Chrome<span class="float-right text-muted">80%</span></p>
-											<div class="progress h-2">
-												<div class="progress-bar bg-primary w-80 " role="progressbar"></div>
-											</div>
-										</div>
-										<div class="mb-5">
-											<p class="mb-2">Firefox<span class="float-right text-muted">70%</span></p>
-											<div class="progress h-2">
-												<div class="progress-bar bg-pink w-70" role="progressbar"></div>
-											</div>
-										</div>
-										<div class="mb-5">
-											<p class="mb-2">Safari<span class="float-right text-muted">80%</span></p>
-											<div class="progress h-2">
-												<div class="progress-bar bg-warning w-65" role="progressbar"></div>
-											</div>
-										</div>
-										<div class="mb-5">
-											<p class="mb-2">Opera<span class="float-right text-muted">80%</span></p>
-											<div class="progress h-2">
-												<div class="progress-bar bg-danger w-80" role="progressbar"></div>
-											</div>
-										</div>
-										<div class="mb-5">
-											<p class="mb-2">UC Browser<span class="float-right text-muted">70%</span></p>
-											<div class="progress h-2">
-												<div class="progress-bar bg-success w-60" role="progressbar"></div>
-											</div>
-										</div>
-										<div class="mb-0">
-											<p class="mb-2">Internet Explore<span class="float-right text-muted">50%</span></p>
-											<div class="progress h-2">
-												<div class="progress-bar bg-info w-50" role="progressbar"></div>
-											</div>
-										</div>
-									</div>
-								</div>
-							</div><!-- COL END -->
-							<div class="col-sm-12 col-md-12 col-lg-12 col-xl-4">
-								<div class="card">
-									<div class="card-header">
-										<h3 class="card-title">Production Status</h3>
-									</div>
-									<div class="p-4 border-bottom">
-										<p class="mb-2 text-muted">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod.</p>
-										<div class="row">
-											<div class="col">
-												<p class="mb-1">Production Volume</p>
-												<h3 class="mb-0 number-font">57,865</h3>
-											</div>
-											<div class="col mb-0">
-												<div id="spark1" class="chart-dropshadow-info"></div>
-											</div>
-										</div>
-									</div>
-									<div class="p-4 border-bottom">
-										<div class="row">
-											<div class="col">
-												<p class="mb-1">Sales Revenue</p>
-												<h3 class="mb-0 number-font">22,345</h3>
-											</div>
-											<div class="col mb-0">
-												<div class="chart-wrapper">
-													<div id="spark2" class="chart-dropshadow-secondary"></div>
-												</div>
-											</div>
-										</div>
-									</div>
-									<div class="p-4 border-bottom">
-										<div class="row">
-											<div class="col">
-												<p class="mb-1">Total Orders</p>
-												<h3 class="mb-0 number-font">52,156</h3>
-											</div>
-											<div class="col mb-0">
-												<div id="spark3" class="chart-dropshadow-success"></div>
-											</div>
-										</div>
-									</div>
-									<div class="p-4">
-										<div class="row">
-											<div class="col">
-												<p class="mb-1">Total profit</p>
-												<h3 class="mb-0 number-font">34,673</h3>
-											</div>
-											<div class="col mb-0">
-												<div id="spark4" class="chart-dropshadow-danger"></div>
-											</div>
-										</div>
-									</div>
-								</div>
-							</div><!-- COL END -->
-							<div class="col-sm-12 col-md-12 col-lg-12 col-xl-4">
-								<div class="card">
-									<div class="card-header">
-										<h3 class="card-title">Total Sales By Org Unit</h3>
-									</div>
-									<div class="card-body">
-										<small class="text-muted">Total Sales</small>
-										<h2 class="number-font">69,453</h2>
-										<div class="progress grouped h-3">
-											<div class="progress-bar w-25 bg-primary " role="progressbar"></div>
-											<div class="progress-bar w-30 bg-danger" role="progressbar"></div>
-											<div class="progress-bar w-20 bg-warning" role="progressbar"></div>
-										</div>
-										<div class="row mt-3 pt-3">
-											<div class="col border-right">
-												<p class=" number-font1 mb-0"><span class="dot-label bg-primary"></span>Sales</p>
-												<h5 class="mt-2 font-weight-semibold mb-0">25%</h5>
-											</div>
-											<div class="col  border-right">
-												<p class=" number-font mb-0"><span class="dot-label bg-danger"></span>Marketing</p>
-												<h5 class="mt-2 font-weight-semibold mb-0">30%</h5>
-											</div>
-											<div class="col">
-												<p class="number-font1 mb-0"><span class="dot-label bg-warning"></span>Finance</p>
-												<h5 class="mt-2 font-weight-semibold mb-0">20%</h5>
-											</div>
-										</div>
-									</div>
-								</div>
-								<div class="card">
-									<div class="card-body">
-										<div class="align-items-center">
-											<h4 class="">Total Expenses</h4>
-										</div>
-										<div class="row d-flex align-items-center mt-3">
-											<div class="col">
-												<h6 class="text-muted mb-2">Monthly</h6>
-												<h2 class="d-flex mb-0 number-font">$35487<small><i class="fe fe-arrow-up text-success"></i></small></h2>
-											</div>
-											<div class="col text-right">
-												<div class="">
-													<span class="sparkline_bar1 chart-dropshadow-secondary"></span>
-												</div>
-											</div>
-										</div>
-									</div>
-								</div>
-							</div><!-- COL END -->
-						</div>
-						<!-- ROW-3 END -->
-
-						<!-- ROW-4 -->
-						<div class="row">
-							<div class="col-sm-12 col-md-12 col-lg-12 col-xl-4">
-								<div class="card overflow-hidden">
-									<div class="card-body pb-0">
-										<div class="">
-											<div class="d-flex">
-												<h6 class="mb-3">Members Online</h6>
-												<div class="ml-auto">
-													<a class="btn btn-light btn-sm btn-icon mr-2" href="#"><i class="fa fa-chevron-left"></i></a>
-													<a class="btn btn-light btn-sm btn-icon" href="#"><i class="fa fa-chevron-right"></i></a>
-												</div>
-											</div>
-											<h3 class="number-font mb-1">2356</h3>
-											<span class="text-success"><i class="zmdi zmdi-long-arrow-up zmdi-hc-lg mr-2"></i><span>+24%</span></span><span class="text-muted ml-2">From Last Month</span>
-											<p class="mb-0 mt-2 text-muted">Lorem ipsum dolor sit amet odio consectetur accusamus .</p>
-										</div>
-									</div>
-									<div class="chart-wrapper">
-										<canvas id="widgetChart1" class="chart-dropshadow-info"></canvas>
-									</div>
-								</div>
-							</div>
-							<div class="col-sm-12 col-md-12 col-lg-12 col-xl-4">
-								<div class="card overflow-hidden">
-									<div class="card-body pb-0">
-										<div class="">
-											<div class="d-flex">
-												<h6 class="mb-3">Current Serverload</h6>
-												<div class="ml-auto">
-													<a class="btn btn-light btn-sm btn-icon mr-2" href="#"><i class="fa fa-chevron-left"></i></a>
-													<a class="btn btn-light btn-sm btn-icon" href="#"><i class="fa fa-chevron-right"></i></a>
-												</div>
-											</div>
-											<h3 class="number-font mb-1">4321</h3>
-											<span class="text-success"><i class="zmdi zmdi-long-arrow-up zmdi-hc-lg mr-2"></i><span>90.5%</span></span><span class="text-muted ml-2">From Last Month</span>
-											<p class="mb-0 mt-2 text-muted">Lorem ipsum dolor sit amet odio consectetur accusamus .</p>
-										</div>
-									</div>
-									<div class="chart-wrapper">
-										<canvas id="widgetChart2" class="chart-dropshadow-secondary"></canvas>
-									</div>
-								</div>
-							</div>
-							<div class="col-sm-12 col-md-12 col-lg-12 col-xl-4">
-								<div class="card overflow-hidden">
-									<div class="card-body pb-0">
-										<div class="">
-											<div class="d-flex">
-												<h6 class="mb-3">To Day's Revenue</h6>
-												<div class="ml-auto">
-													<a class="btn btn-light btn-sm btn-icon mr-2" href="#"><i class="fa fa-chevron-left"></i></a>
-													<a class="btn btn-light btn-sm btn-icon" href="#"><i class="fa fa-chevron-right"></i></a>
-												</div>
-											</div>
-											<h3 class="number-font mb-1">8963</h3>
-											<span class="text-success"><i class="zmdi zmdi-long-arrow-up zmdi-hc-lg mr-2"></i><span>20.8%</span></span><span class="text-muted ml-2">From Last Month</span>
-											<p class="mb-0 mt-2 text-muted">Lorem ipsum dolor sit amet odio consectetur accusamus .</p>
-										</div>
-									</div>
-									<div class="chart-wrapper">
-										<canvas id="widgetChart3" class="chart-dropshadow-success"></canvas>
-									</div>
-								</div>
-							</div>
-						</div>
-						<!-- ROW-4 END -->
-
-						<!-- ROW-5 -->
-						<div class="row">
-							<div class="col-12 col-sm-12">
-								<div class="card ">
-									<div class="card-header">
-										<h3 class="card-title mb-0">Product Sales</h3>
-									</div>
-									<div class="card-body">
-										<div class="grid-margin">
-											<div class="">
-												<div class="table-responsive">
-													<table class="table card-table border table-vcenter text-nowrap align-items-center">
-														<thead class="">
-															<tr>
-																<th>Product</th>
-																<th>Sold</th>
-																<th>Record point</th>
-																<th>Stock</th>
-																<th>Amount</th>
-																<th>Stock Status</th>
-															</tr>
-														</thead>
-														<tbody>
-															<tr>
-																<td class="text-sm font-weight-600">Women Wallet-E32N</td>
-																<td>18</td>
-																<td>05</td>
-																<td>112</td>
-																<td>$ 2,356</td>
-																<td><i class="fa fa-check mr-1 text-success"></i> In Stock</td>
-															</tr>
-															<tr>
-																<td class="text-sm font-weight-600">Phone-345S</td>
-																<td>10</td>
-																<td>04</td>
-																<td>210</td>
-																<td>$ 3,522</td>
-																<td><i class="fa fa-check text-success"></i> In Stock</td>
-															</tr>
-															<tr>
-																<td class="text-sm font-weight-600">Doll-Elephant</td>
-																<td>15</td>
-																<td>05</td>
-																<td>215</td>
-																<td>$ 5,362</td>
-																<td><i class="fa fa-check text-success"></i> In Stock</td>
-															</tr>
-															<tr>
-																<td class="text-sm font-weight-600">Women-Kurtis</td>
-																<td>21</td>
-																<td>07</td>
-																<td>102</td>
-																<td>$ 1,326</td>
-																<td><i class="fa fa-check text-success"></i> In Stock</td>
-															</tr>
-															<tr>
-																<td class="text-sm font-weight-600">Mens Shoes</td>
-																<td>34</td>
-																<td>10</td>
-																<td>325</td>
-																<td>$ 5,234</td>
-																<td><i class="fa fa-check text-success"></i> In Stock</td>
-															</tr>
-															<tr>
-																<td class="text-sm font-weight-600">Mens Gold Metal Watch</td>
-																<td>11</td>
-																<td>04</td>
-																<td>0</td>
-																<td>$ 3,256</td>
-																<td>
-																	<i class="fa fa-exclamation-triangle text-warning"></i>  Out of stock
-																</td>
-															</tr>
-															<tr>
-																<td class="text-sm font-weight-600">Laptop</td>
-																<td>60</td>
-																<td>10</td>
-																<td>0</td>
-																<td>$ 7,652</td>
-																<td>
-																	<i class="fa fa-exclamation-triangle text-danger"></i> Out of stock
-																</td>
-															</tr>
-														</tbody>
-													</table>
-												</div>
-											</div>
-										</div>
-									</div>
-								</div>
-							</div><!-- COL END -->
-						</div><!-- ROW-5 END -->
-					</div>
-				</div>
-				<!-- CONTAINER END -->
-            </div>
-           
-  
-            <!-- FOOTER -->
-			<footer class="footer">
-				<div class="container">
-					<div class="row align-items-center flex-row-reverse">
-						<div class="col-md-12 col-sm-12 text-center">
-							Copyright © 2019 <a href="#">Volgh</a>. Designed by <a href="#"> Spruko Technologies Pvt.Ltd </a> All rights reserved.
-						</div>
-					</div>
-				</div>
-			</footer>
-<!-- FOOTER END -->   
-        </div>
-        </div>
-            <!-- BACK-TO-TOP -->
-		<a href="#top" id="back-to-top"><i class="fa fa-angle-up"></i></a>
-		<!-- JQUERY JS -->
-		<script src="{{ asset('assets/js/jquery-3.4.1.min.js') }}"></script>
-
-		<!-- BOOTSTRAP JS -->
-		<script src="{{ asset('assets/plugins/bootstrap/js/bootstrap.bundle.min.js') }}"></script>
-		<script src="{{ asset('assets/plugins/bootstrap/js/popper.min.js') }}"></script>
-
-		<!-- SPARKLINE JS-->
-		<script src="{{ asset('assets/js/jquery.sparkline.min.js') }}"></script>
-
-		<!-- CHART-CIRCLE JS -->
-		<script src="{{ asset('assets/js/circle-progress.min.js') }}"></script>
-
-		<!-- RATING STAR JS-->
-		<script src="{{ asset('assets/plugins/rating/jquery.rating-stars.js') }}"></script>
-
-		<!-- C3.JS CHART JS -->
-		<script src="{{ asset('assets/plugins/charts-c3/d3.v5.min.js') }}"></script>
-		<script src="{{ asset('assets/plugins/charts-c3/c3-chart.js') }}"></script>
-
-		<!-- INPUT MASK JS-->
-		<script src="{{ asset('assets/plugins/input-mask/jquery.mask.min.js') }}"></script>
-
-		<!-- SIDE-MENU JS-->
-		<script src="{{ asset('assets/plugins/sidemenu/sidemenu.js') }}"></script>
-
-        <!-- CUSTOM SCROLLBAR JS-->
-		<script src="{{ asset('assets/plugins/scroll-bar/jquery.mCustomScrollbar.concat.min.js') }}"></script>
-		
-		<script src="{{ asset('assets/js/index1.js') }}"></script>
-<script src="{{ asset('assets/plugins/chart/Chart.bundle.js') }}"></script>
-<script src="{{ asset('assets/plugins/chart/utils.js') }}"></script>
-<script src="{{ asset('assets/js/apexcharts.js') }}"></script>
-<script src="{{ asset('assets/plugins/peitychart/jquery.peity.min.js') }}"></script>
-<script src="{{ asset('assets/plugins/peitychart/peitychart.init.js') }}"></script>
-<script src="{{asset('assets/plugins/echarts/echarts.js') }}"></script>
-
-		<!-- SIDEBAR JS -->
-		<script src="{{ asset('assets/plugins/sidebar/sidebar.js') }}"></script>
-		
-		<!-- P-scroll bar js-->
-		<script src="{{ asset('assets/plugins/p-scroll/perfect-scrollbar.js') }}"></script>
-		<script src="{{ asset('assets/plugins/p-scroll/p-scroll.js') }}"></script>
-		
-		<!-- SWITCHER JS -->
-		<script src="{{ asset('assets/switcher/js/switcher.js') }}"></script>
-		
-		<!--CUSTOM JS -->
-		<script src="{{ asset('assets/js/custom.js') }}"></script>
-
-
-
-  
-    </body>
-
-<!-- Mirrored from laravel.spruko.com/volgh/leftmenu_icon_light/index by HTTrack Website Copier/3.x [XR&CO'2014], Sat, 19 Dec 2020 16:20:50 GMT -->
-</html>
